backend/pyproject.toml
[project]
name = "teststructure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt<4.1.0",
    "fastapi>=0.128.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
